# apps/credential-showcase-api-server/src/controllers/IssuanceScenarioController.ts
import { Body, Delete, Get, HttpCode, JsonController, OnUndefined, Param, Post, Put } from 'routing-controllers'
import { Service } from 'typedi'
import ScenarioService from '../services/ScenarioService'
import {
  IssuanceScenarioResponse,
  IssuanceScenarioResponseFromJSONTyped,
  IssuanceScenariosResponse,
  IssuanceScenariosResponseFromJSONTyped,
  StepsResponse,
  StepsResponseFromJSONTyped,
  StepResponse,
  StepResponseFromJSONTyped,
  StepRequest,
  StepRequestToJSONTyped,
  StepActionsResponse,
  StepActionsResponseFromJSONTyped,
  StepActionResponse,
  StepActionResponseFromJSONTyped,
  StepActionRequest,
  StepActionRequestToJSONTyped,
  IssuanceScenarioRequest,
  IssuanceScenarioRequestToJSONTyped,
} from 'credential-showcase-openapi'
import { issuanceScenarioDTOFrom, stepDTOFrom } from '../utils/mappers'
import { ScenarioType } from '../types'

@JsonController('/scenarios/issuances')
@Service()
class IssuanceScenarioController {
  constructor(private scenarioService: ScenarioService) {}

  @Get('/')
  public async getAllIssuanceScenarios(): Promise<IssuanceScenariosResponse> {
    const result = await this.scenarioService.getScenarios({ filter: { scenarioType: ScenarioType.ISSUANCE } })
    const issuanceScenarios = result.map((issuanceScenario) => issuanceScenarioDTOFrom(issuanceScenario))
    return IssuanceScenariosResponseFromJSONTyped({ issuanceScenarios }, false)
  }

  @Get('/:slug')
  public async getOneIssuanceScenario(@Param('slug') slug: string): Promise<IssuanceScenarioResponse> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    const result = await this.scenarioService.getScenario(issuanceScenarioId)
    return IssuanceScenarioResponseFromJSONTyped({ issuanceScenario: issuanceScenarioDTOFrom(result) }, false)
  }

  @HttpCode(201)
  @Post('/')
  public async postIssuanceScenario(@Body() issuanceScenarioRequest: IssuanceScenarioRequest): Promise<IssuanceScenarioResponse> {
    const result = await this.scenarioService.createScenario(IssuanceScenarioRequestToJSONTyped(issuanceScenarioRequest))
    return IssuanceScenarioResponseFromJSONTyped({ issuanceScenario: issuanceScenarioDTOFrom(result) }, false)
  }

  @Put('/:slug')
  public async putIssuanceScenario(
    @Param('slug') slug: string,
    @Body() issuanceScenarioRequest: IssuanceScenarioRequest,
  ): Promise<IssuanceScenarioResponse> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    const result = await this.scenarioService.updateScenario(issuanceScenarioId, IssuanceScenarioRequestToJSONTyped(issuanceScenarioRequest))
    return IssuanceScenarioResponseFromJSONTyped({ issuanceScenario: issuanceScenarioDTOFrom(result) }, false)
  }

  @OnUndefined(204)
  @Delete('/:slug')
  public async deleteIssuanceScenario(@Param('slug') slug: string): Promise<void> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    return await this.scenarioService.deleteScenario(issuanceScenarioId)
  }

  @Get('/:slug/steps')
  public async getAllSteps(@Param('slug') slug: string): Promise<StepsResponse> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    const result = await this.scenarioService.getScenarioSteps(issuanceScenarioId)
    const steps = result.map((step) => stepDTOFrom(step))
    return StepsResponseFromJSONTyped({ steps }, false)
  }

  @Get('/:slug/steps/:stepId')
  public async getOneIssuanceScenarioStep(
    @Param('slug') slug: string,
    @Param('stepId') stepId: string,
  ): Promise<StepResponse> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    const result = await this.scenarioService.getScenarioStep(issuanceScenarioId, stepId)
    return StepResponseFromJSONTyped({ step: stepDTOFrom(result) }, false)
  }

  @HttpCode(201)
  @Post('/:slug/steps')
  public async postIssuanceScenarioStep(
    @Param('slug') slug: string,
    @Body() stepRequest: StepRequest,
  ): Promise<StepResponse> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    const result = await this.scenarioService.createScenarioStep(issuanceScenarioId, StepRequestToJSONTyped(stepRequest))
    return StepResponseFromJSONTyped({ step: stepDTOFrom(result) }, false)
  }

  @Put('/:slug/steps/:stepId')
  public async putIssuanceScenarioStep(
    @Param('slug') slug: string,
    @Param('stepId') stepId: string,
    @Body() stepRequest: StepRequest,
  ): Promise<StepResponse> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    const result = await this.scenarioService.updateScenarioStep(issuanceScenarioId, stepId, StepRequestToJSONTyped(stepRequest))
    return StepResponseFromJSONTyped({ step: stepDTOFrom(result) }, false)
  }

  @OnUndefined(204)
  @Delete('/:slug/steps/:stepId')
  public async deleteIssuanceScenarioStep(@Param('slug') slug: string, @Param('stepId') stepId: string): Promise<void> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    return this.scenarioService.deleteScenarioStep(issuanceScenarioId, stepId)
  }

  @Get('/:slug/steps/:stepId/actions')
  public async getAllIssuanceScenarioStepActions(
    @Param('slug') slug: string,
    @Param('stepId') stepId: string,
  ): Promise<StepActionsResponse> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    const result = await this.scenarioService.getScenarioStepActions(issuanceScenarioId, stepId)
    const actions = result.map((action) => action)
    return StepActionsResponseFromJSONTyped({ actions }, false)
  }

  @Get('/:slug/steps/:stepId/actions/:actionId')
  public async getOneIssuanceScenarioStepAction(
    @Param('slug') slug: string,
    @Param('stepId') stepId: string,
    @Param('actionId') actionId: string,
  ): Promise<StepActionResponse> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    const result = await this.scenarioService.getScenarioStepAction(issuanceScenarioId, stepId, actionId)
    return StepActionResponseFromJSONTyped({ action: result }, false)
  }

  @HttpCode(201)
  @Post('/:slug/steps/:stepId/actions')
  public async postIssuanceScenarioStepAction(
    @Param('slug') slug: string,
    @Param('stepId') stepId: string,
    @Body() actionRequest: StepActionRequest,
  ): Promise<StepActionResponse> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    const result = await this.scenarioService.createScenarioStepAction(issuanceScenarioId, stepId, StepActionRequestToJSONTyped(actionRequest))
    return StepActionResponseFromJSONTyped({ action: result }, false)
  }

  @Put('/:slug/steps/:stepId/actions/:actionId')
  public async putIssuanceScenarioStepAction(
    @Param('slug') slug: string,
    @Param('stepId') stepId: string,
    @Param('actionId') actionId: string,
    @Body() actionRequest: StepActionRequest,
  ): Promise<StepActionResponse> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    const result = await this.scenarioService.updateScenarioStepAction(
      issuanceScenarioId,
      stepId,
      actionId,
      StepActionRequestToJSONTyped(actionRequest),
    )
    return StepActionResponseFromJSONTyped({ action: result }, false)
  }

  @OnUndefined(204)
  @Delete('/:slug/steps/:stepId/actions/:actionId')
  public async deleteIssuanceScenarioStepAction(
    @Param('slug') slug: string,
    @Param('stepId') stepId: string,
    @Param('actionId') actionId: string,
  ): Promise<void> {
    const issuanceScenarioId = await this.scenarioService.getIdBySlug(slug)
    return this.scenarioService.deleteScenarioStepAction(issuanceScenarioId, stepId, actionId)
  }
}

export default IssuanceScenarioController